assignment/requirements.txt
selenium>=4.0.0
beautifulsoup4
requests
python-dotenv
webdriver-manager
deep-translator
